+++ src/main/java/com/erayt/cuda/GpuInterface.java
@@ -3,5 +3,5 @@ package com.erayt.cuda;
 public class GpuInterface {
     public static native void cudaInit();
 
-    public static native float[] run(int algoId, float[] args);
+    public static native double[] run(int algoId, double[] args);
 }
