+++ src/main/java/com/erayt/cuda/Main.java
@@ -26,15 +26,15 @@ public class Main {
         System.out.println("Monte Carlo");
         // === Monte Carlo 模拟
         int numPaths = 10_000_000;
-        float S0 = 100f;
-        float K = 100f;
-        float r = 0.05f;
-        float sigma = 0.2f;
-        float T = 1f;
+        double S0 = 100;
+        double K = 100;
+        double r = 0.05;
+        double sigma = 0.2;
+        double T = 1;
 
         // === CPU Monte Carlo
         long start = System.currentTimeMillis();
-        float price = europeanCallPrice(numPaths, S0, K, r, sigma, T);
+        double price = europeanCallPrice(numPaths, S0, K, r, sigma, T);
         long end = System.currentTimeMillis();
         System.out.println("CPU Time: " + (end - start) + " ms");
         System.out.println("CPU Monte Carlo price: " + price);
@@ -42,7 +42,7 @@ public class Main {
         // === GPU Monte Carlo
         start = System.currentTimeMillis();
         MonteAlgo algo = MonteAlgo.of(numPaths, S0, K, r, sigma, T);
-        float[] discounted = GpuInterface.run(algo.getId(), algo.toArgs());
+        double[] discounted = GpuInterface.run(algo.getId(), algo.toArgs());
         end = System.currentTimeMillis();
         System.out.println("GPU Time: " + (end - start) + " ms");
         System.out.println("GPU Monte Carlo price: " + discounted[0]);
@@ -54,20 +54,20 @@ public class Main {
         System.out.println("Var");
         // === Var
         int n = 1_000_000;
-        float[] returns = new float[n];
+        double[] returns = new double[n];
         Random rand = new Random(42);
 
         for (int i = 0; i < n; i++) {
             // 正态分布模拟历史收益（均值0.001，标准差0.02）
-            returns[i] = (float) (0.001 + rand.nextGaussian() * 0.02);
+            returns[i] = 0.001 + rand.nextGaussian() * 0.02;
         }
 
-        float confidence = 0.99f;
-        float holding = 1.0f;
+        double confidence = 0.99;
+        double holding = 1.0;
 
         // === CPU Var
         long start = System.currentTimeMillis();
-        float javaVaR = computeVaR(returns, confidence, holding);
+        double javaVaR = computeVaR(returns, confidence, holding);
         long end = System.currentTimeMillis();
         System.out.println("CPU Time: " + (end - start) + " ms");
         System.out.println("Java VaR = " + javaVaR);
@@ -75,7 +75,7 @@ public class Main {
         // === GPU Var
         start = System.currentTimeMillis();
         VarAlgo varAlgo = VarAlgo.of(n, returns, confidence, holding);
-        float[] var = GpuInterface.run(varAlgo.getId(), varAlgo.toArgs());
+        double[] var = GpuInterface.run(varAlgo.getId(), varAlgo.toArgs());
         end = System.currentTimeMillis();
         System.out.println("GPU Time: " + (end - start) + " ms");
         System.out.println("GPU VaR = " + var[0]);
@@ -87,29 +87,29 @@ public class Main {
         System.out.println("Vol Surface");
         int N = 1_000_000;
         int count = 5;
-        float[] S = new float[N];
-        float[] K = new float[N];
-        float[] T = new float[N];
-        float[] r = new float[N];
-        float[] P = new float[N];
+        double[] S = new double[N];
+        double[] K = new double[N];
+        double[] T = new double[N];
+        double[] r = new double[N];
+        double[] P = new double[N];
 
         Random rand = new Random(42);
         for (int i = 0; i < N; i++) {
-            S[i] = 90 + rand.nextFloat() * 20; // spot price: 90 ~ 110
-            K[i] = 90 + rand.nextFloat() * 20; // strike price: 90 ~ 110
-            T[i] = 0.1f + rand.nextFloat(); // maturity: 0.1 ~ 1.1 years
-            r[i] = 0.01f; // interest rate
+            S[i] = 90 + rand.nextDouble() * 20; // spot price: 90 ~ 110
+            K[i] = 90 + rand.nextDouble() * 20; // strike price: 90 ~ 110
+            T[i] = 0.1 + rand.nextDouble(); // maturity: 0.1 ~ 1.1 years
+            r[i] = 0.01; // interest rate
         }
 
         // 用固定sigma算出市场价格
-        float trueSigma = 0.2f;
+        double trueSigma = 0.2;
         for (int i = 0; i < N; i++) {
-            P[i] = blackScholesCall(S[i], K[i], T[i], r[i], 0.2f) + (float) (Math.random() - 0.5) * 0.5f;
+            P[i] = blackScholesCall(S[i], K[i], T[i], r[i], 0.2) + (Math.random() - 0.5) * 0.5;
         }
 
         // CPU计算隐含波动率
         long start = System.currentTimeMillis();
-        float[] cpuVol = new float[N];
+        double[] cpuVol = new double[N];
         for (int i = 0; i < N; i++) {
             cpuVol[i] = impliedVol(S[i], K[i], T[i], r[i], P[i]);
         }
@@ -138,7 +138,7 @@ public class Main {
         VolSurface volSurface = VolSurface.of(N, S, K, T, r, P);
         // === GPU
         start = System.currentTimeMillis();
-        float[] gpuVol = GpuInterface.run(volSurface.getId(), volSurface.toArgs());
+        double[] gpuVol = GpuInterface.run(volSurface.getId(), volSurface.toArgs());
         end = System.currentTimeMillis();
         System.out.println("GPU total time: " + (end - start) + " ms");
 
@@ -146,62 +146,63 @@ public class Main {
             System.out.printf("S=%.2f K=%.2f T=%.2f -> IV=%.4f (true=%.4f)\n",
                     S[i], K[i], T[i], gpuVol[i], trueSigma);
         }
+
     }
 
-    static float europeanCallPrice(
+    static double europeanCallPrice(
             int numPaths,
-            float s0,
-            float k,
-            float r,
-            float sigma,
-            float t) {
+            double s0,
+            double k,
+            double r,
+            double sigma,
+            double t) {
         Random rand = new Random();
-        float sumPayoff = 0.0f;
+        double sumPayoff = 0.0;
 
         for (int i = 0; i < numPaths; i++) {
             // 生成一个标准正态分布随机数 Z
-            float z = (float) rand.nextGaussian();
+            double z = rand.nextGaussian();
 
             // 计算终止价格 S_T
-            float st = (float) (s0 * Math.exp((r - 0.5 * sigma * sigma) * t + sigma * Math.sqrt(t) * z));
+            double st = s0 * Math.exp((r - 0.5 * sigma * sigma) * t + sigma * Math.sqrt(t) * z);
 
             // 看涨期权 payoff
-            float payoff = Math.max(st - k, 0.0f);
+            double payoff = Math.max(st - k, 0.0);
 
             sumPayoff += payoff;
         }
 
-        float meanPayoff = sumPayoff / numPaths;
+        double meanPayoff = sumPayoff / numPaths;
 
         // 折现回当前价格
-        return (float) (Math.exp(-r * t) * meanPayoff);
+        return Math.exp(-r * t) * meanPayoff;
     }
 
-    static float computeVaR(float[] returns, float confidence, float holding) {
+    static double computeVaR(double[] returns, double confidence, double holding) {
         int n = returns.length;
-        float[] losses = new float[n];
+        double[] losses = new double[n];
         Random rand = new Random();
 
-        float Z = (float) rand.nextGaussian();
+        double Z = rand.nextGaussian();
 
         for (int i = 0; i < n; i++) {
-            losses[i] = holding - returns[i] * (1 + Z * 0.01f);
+            losses[i] = holding - returns[i] * (1 + Z * 0.01);
         }
 
         java.util.Arrays.sort(losses);
 
-        int index = (int) ((1.0f - confidence) * n);
+        int index = (int) ((1.0 - confidence) * n);
         return losses[index];
     }
 
-    static float impliedVol(float S, float K, float T, float r, float marketPrice) {
-        float sigma = 0.2f;
+    static double impliedVol(double S, double K, double T, double r, double marketPrice) {
+        double sigma = 0.2;
         for (int i = 0; i < 20; i++) {
-            float d1 = (float) ((Math.log(S / K) + (r + 0.5 * sigma * sigma) * T) / (sigma * Math.sqrt(T)));
-            float d2 = d1 - sigma * (float) Math.sqrt(T);
-            float call = (float) (S * normCdf(d1) - K * Math.exp(-r * T) * normCdf(d2));
-            float vega = (float) (S * Math.sqrt(T) * normPdf(d1));
-            float diff = call - marketPrice;
+            double d1 = (Math.log(S / K) + (r + 0.5 * sigma * sigma) * T) / (sigma * Math.sqrt(T));
+            double d2 = d1 - sigma * Math.sqrt(T);
+            double call = S * normCdf(d1) - K * Math.exp(-r * T) * normCdf(d2);
+            double vega = S * Math.sqrt(T) * normPdf(d1);
+            double diff = call - marketPrice;
             if (Math.abs(diff) < 1e-6)
                 break;
             sigma -= diff / vega;
@@ -209,18 +210,18 @@ public class Main {
         return sigma;
     }
 
-    static float blackScholesCall(float S, float K, float T, float r, float sigma) {
-        float d1 = (float) ((Math.log(S / K) + (r + 0.5 * sigma * sigma) * T) / (sigma * Math.sqrt(T)));
-        float d2 = d1 - sigma * (float) Math.sqrt(T);
-        return S * normCdf(d1) - K * (float) Math.exp(-r * T) * normCdf(d2);
+    static double blackScholesCall(double S, double K, double T, double r, double sigma) {
+        double d1 = (Math.log(S / K) + (r + 0.5 * sigma * sigma) * T) / (sigma * Math.sqrt(T));
+        double d2 = d1 - sigma * Math.sqrt(T);
+        return S * normCdf(d1) - K * Math.exp(-r * T) * normCdf(d2);
     }
 
-    static float normPdf(float x) {
-        return (float) (1.0 / Math.sqrt(2 * Math.PI) * Math.exp(-0.5 * x * x));
+    static double normPdf(double x) {
+        return 1.0 / Math.sqrt(2 * Math.PI) * Math.exp(-0.5 * x * x);
     }
 
-    static float normCdf(float x) {
-        return (float) (0.5 * (1 + erf(x / Math.sqrt(2))));
+    static double normCdf(double x) {
+        return 0.5 * (1 + erf(x / Math.sqrt(2)));
     }
 
     // 近似误差函数
